Consolidate Gemma4 numeric checker
