+++ models/gemma4_e2b/gemma4_e2b_numeric.py
@@ -1,452 +1,558 @@
 #!/usr/bin/env python3
-"""Gemma4 E2B VISION-encoder numeric verification: host oracle vs FPGA.
+"""LM and optional vision numeric checks for the REFACTOR engine.
 
-Debug tool for the vision encoder ONLY (that is the part under investigation).
+Builds full-precision "hf" and IF4-quantized "hostref" vision references and
+runs the FPGA side through Gemma4_UnifiedEngine. It SNR-compares the
+self._vis_ckpt readbacks at two stages:
 
-The host reference *mimics the hardware*: it runs the HF Gemma4 vision tower with
-every Linear weight round-tripped through IF4 — the same block-quantization the
-FPGA uses for the vision matmuls (VISION_QUANT_PRECISION) — so the quantization
-gap is present in BOTH sides and the residual SNR reflects HARDWARE execution
-error, not the (benign, identical-on-every-HW) quant/algorithm gap.
+  encoder_out   -> reference B  (after all 16 encoder layers, pre-pool)
+  image_features-> reference C  (after pooler + embedding projection)
 
-Each run reports the FPGA readbacks (stashed by _run_vision_encoder_fpga in
-self._vis_ckpt) SNR-compared against BOTH references:
-  * hostref — HF vision tower with IF4-quantized weights (mimics hardware).
-  * hf      — full-precision HF vision tower (ground truth).
+Then checks LM prefill final hidden states and the first decoder token's final
+hidden state, output norm, logits and argmax. The LM HF reference consumes the
+exact FPGA-produced image soft tokens, so vision error does not leak into LM.
 
-  A = patch_embed    [S, H]           (VIS_IO_A)
-  B = encoder_out    [S, H]           (final encoder hidden state, pre-pool)
-  C = image_features [N_soft, 1536]   (pooler + embed_vision projection)
+Interpretation (same as numeric.py):
+  * FPGA vs HOSTREF (both IF4) should be HIGH SNR — a low value is a real kernel
+    bug, not quantization.
+  * FPGA vs HF is lower by the IF4 quantization loss (ground-truth gap).
 
-Uses calculate_snr from user_dma_core, mirroring gemma3_numeric.py.
+Requires params.bin regenerated with the [LM | vision | host] layout.
 
 Usage:
-  python gemma4_e2b_numeric.py --dev xdma1
+  python models/gemma4_e2b/gemma4_e2b_numeric.py            # LM-only, default prompt
+  python models/gemma4_e2b/gemma4_e2b_numeric.py --image    # yosemite + VLM
+  python models/gemma4_e2b/gemma4_e2b_numeric.py --image people.jpg
+  python models/gemma4_e2b/gemma4_e2b_numeric.py --dev xdma0 --cycle 5.042
 """
+import argparse
 import os
 import sys
-import argparse
 
 SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
-sys.path.insert(0, SCRIPT_DIR)                                   # import gemma4_e2b_test
-sys.path.insert(0, os.path.dirname(os.path.dirname(SCRIPT_DIR)))  # repo root (user_dma_core)
+sys.path.insert(0, SCRIPT_DIR)                                   # gemma4_e2b_test / _numeric
+sys.path.insert(0, os.path.dirname(os.path.dirname(SCRIPT_DIR)))  # user_dma_core (repo root)
 
+import user_dma_core
 import torch
 import torch.nn as nn
 from PIL import Image
-
-import user_dma_core
-from user_dma_core import set_dma_device, calculate_snr
-import quant_lib
-import gemma4_e2b_test as g4
-from gemma4_e2b_test import (
-    Gemma4_UnifiedEngine, _ensure_hf_model,
-    VISION_CANONICAL_SIZE, VISION_QUANT_PRECISION,
-    DEFAULT_IMAGE, DEFAULT_AUDIO,
-)
 from transformers import AutoProcessor
+from user_dma_core import calculate_snr, set_dma_device
+import gemma4_e2b_test as g4r
+import gemma4_e2b_lm as g4_lm_module
+import gemma4_e2b_vision as g4_vision_module
+import quant_lib
 
 
-# ---------------------------------------------------------------------------
-# Metrics
-# ---------------------------------------------------------------------------
 def _align(ref, res):
-    """Flatten batch, match row count / shape between host ref and FPGA readback."""
-    r = ref.detach().float()
-    f = res.detach().float()
-    if r.dim() == 3 and r.shape[0] == 1:
-        r = r.squeeze(0)
-    if f.dim() == 3 and f.shape[0] == 1:
-        f = f.squeeze(0)
-    if r.shape != f.shape:
-        r = r.reshape(-1, r.shape[-1])
-        f = f.reshape(-1, f.shape[-1])
-        n = min(r.shape[0], f.shape[0])
-        r, f = r[:n], f[:n]
-    return r, f
+    """Flatten batch dimensions and align host and FPGA rows."""
+    ref = ref.detach().float()
+    res = res.detach().float()
+    if ref.dim() == 3 and ref.shape[0] == 1:
+        ref = ref.squeeze(0)
+    if res.dim() == 3 and res.shape[0] == 1:
+        res = res.squeeze(0)
+    if ref.shape != res.shape:
+        ref = ref.reshape(-1, ref.shape[-1])
+        res = res.reshape(-1, res.shape[-1])
+        rows = min(ref.shape[0], res.shape[0])
+        ref, res = ref[:rows], res[:rows]
+    return ref, res
 
 
 def report(name, ref, res, row_mask=None):
-    """Print SNR / rel_L2 / max|Δ| of FPGA `res` vs host `ref`.
-
-    row_mask (bool over rows) restricts the comparison to real (non-padding)
-    patches so the ~2264 zero padding rows don't dominate the norms.
-    """
-    r, f = _align(ref, res)
-    if row_mask is not None and row_mask.numel() == r.shape[0]:
-        r, f = r[row_mask], f[row_mask]
-    m = torch.isfinite(r).all(dim=-1) & torch.isfinite(f).all(dim=-1)
-    r, f = r[m], f[m]
-    if r.numel() == 0:
+    """Print SNR, relative L2 error, and maximum absolute error."""
+    ref, res = _align(ref, res)
+    if row_mask is not None and row_mask.numel() == ref.shape[0]:
+        ref, res = ref[row_mask], res[row_mask]
+    finite = torch.isfinite(ref).all(dim=-1) & torch.isfinite(res).all(dim=-1)
+    ref, res = ref[finite], res[finite]
+    if ref.numel() == 0:
         print(f"  [numeric] {name:20s} no finite overlap")
-        return float('-inf')
-    snr = float(calculate_snr(r, f))
-    err = float(torch.linalg.vector_norm(f - r))
-    sig = float(torch.linalg.vector_norm(r))
-    rel = err / sig if sig > 0 else float('inf')
-    maxd = float((f - r).abs().max())
-    print(f"  [numeric] {name:20s} SNR={snr:7.2f} dB  rel_L2={rel:.4g}  "
-          f"max|Δ|={maxd:.4g}  rows={r.shape[0]} shape={tuple(f.shape)}")
+        return float("-inf")
+    snr = float(calculate_snr(ref, res))
+    error = float(torch.linalg.vector_norm(res - ref))
+    signal = float(torch.linalg.vector_norm(ref))
+    relative = error / signal if signal > 0 else float("inf")
+    max_delta = float((res - ref).abs().max())
+    print(f"  [numeric] {name:20s} SNR={snr:7.2f} dB  rel_L2={relative:.4g}  "
+          f"max|Δ|={max_delta:.4g}  rows={ref.shape[0]} shape={tuple(res.shape)}")
     return snr
 
 
-# ---------------------------------------------------------------------------
-# Host reference (HF vision tower, optionally IF4-quantized weights)
-# ---------------------------------------------------------------------------
-def quantize_vision_tower_(vt, precision, block=64):
-    """Round-trip every 2D Linear weight in the vision tower through `precision`
-    (mimics the FPGA's quantized vision weights). Modifies vt in place; returns
-    the number of weights quantized."""
-    n = 0
+def quantize_vision_tower_(vision_tower, precision, block=64):
+    """Round-trip every vision Linear weight through the FPGA precision."""
+    count = 0
     with torch.no_grad():
-        for mod in vt.modules():
-            w = getattr(mod, "weight", None)
-            if isinstance(mod, nn.Linear) and w is not None and w.dim() == 2 and w.shape[1] % block == 0:
-                wq = w.detach().cpu().to(torch.bfloat16)
-                N, K = wq.shape
-                data, scale = quant_lib.quantize(precision, wq, block)
-                deq = quant_lib.dequant(precision, data, scale, N, K, block)
-                w.data.copy_(deq.to(w.dtype).to(w.device))
-                n += 1
-    return n
+        for module in vision_tower.modules():
+            weight = getattr(module, "weight", None)
+            if (isinstance(module, nn.Linear) and weight is not None
+                    and weight.dim() == 2 and weight.shape[1] % block == 0):
+                quant_input = weight.detach().cpu().to(torch.bfloat16)
+                n, k = quant_input.shape
+                data, scale = quant_lib.quantize(precision, quant_input, block)
+                dequantized = quant_lib.dequant(
+                    precision, data, scale, n, k, block)
+                weight.data.copy_(dequantized.to(weight.dtype).to(weight.device))
+                count += 1
+    return count
 
 
-def _tensor(o):
-    t = getattr(o, "last_hidden_state", None)
-    if t is None:
-        t = o[0] if isinstance(o, (tuple, list)) else o
-    return t
+def _output_tensor(output):
+    """Extract a tensor from HF tensor/tuple/ModelOutput hook results."""
+    if hasattr(output, "last_hidden_state"):
+        return output.last_hidden_state
+    if isinstance(output, (tuple, list)):
+        return output[0]
+    return output
 
 
-def _add_l0_attention_layout_refs(caps, vt, pixel_position_ids):
-    """Add layer-0 attention internals in the FPGA's per-head RoPE layout.
-
-    The FPGA permutes Q/K/V channels within each head before RoPE and permutes
-    o_proj input columns to consume that layout. These refs are therefore not
-    the vanilla HF layout; they match the checkpoint buffers captured from DRAM.
-    """
-    try:
-        from transformers.models.gemma4.modeling_gemma4 import apply_multidimensional_rope
-        L0 = vt.encoder.layers[0]
-        attn = L0.self_attn
-        hidden = caps.get("L0_before")
-        if hidden is None:
-            return
-        H = hidden.shape[-1]
-        HD = attn.head_dim
-        NH = H // HD
-        S = hidden.shape[-2]
-        aligned_S = ((S + 63) // 64) * 64
-        perm = torch.cat([
-            torch.arange(0, 16),
-            torch.arange(32, 48),
-            torch.arange(16, 32),
-            torch.arange(48, 64),
-        ]).to(hidden.device)
-
-        def _fpga_interleaved(x):
-            # [1,S,NH,HD] standard HF order -> [S,NH*HD] FPGA per-head perm order.
-            return x.squeeze(0)[:, :, perm].reshape(S, NH * HD).detach()
-
-        def _head_major(x_interleaved, prescale=1.0):
-            x = x_interleaved.reshape(S, NH, HD).transpose(0, 1).contiguous()
-            out = torch.zeros((NH, aligned_S, HD), dtype=x.dtype, device=x.device)
-            out[:, :S, :] = x * prescale
-            return out.reshape(NH * aligned_S, HD).detach()
-
-        cos, sin = vt.encoder.rotary_emb(caps["A"], pixel_position_ids)
-        hidden_shape = (*hidden.shape[:-1], NH, HD)
-        q_proj = attn.q_proj(hidden).view(hidden_shape)
-        k_proj = attn.k_proj(hidden).view(hidden_shape)
-        v_proj = attn.v_proj(hidden).view(hidden_shape)
-        q_norm = attn.q_norm(q_proj)
-        k_norm = attn.k_norm(k_proj)
-        v_norm = attn.v_norm(v_proj)
-        rope_q = apply_multidimensional_rope(q_norm, cos, sin, pixel_position_ids)
-        rope_k = apply_multidimensional_rope(k_norm, cos, sin, pixel_position_ids)
-
-        q_proj_i = _fpga_interleaved(q_proj)
-        k_proj_i = _fpga_interleaved(k_proj)
-        v_proj_i = _fpga_interleaved(v_proj)
-        q_norm_i = _fpga_interleaved(q_norm)
-        k_norm_i = _fpga_interleaved(k_norm)
-        v_norm_i = _fpga_interleaved(v_norm)
-        rope_q_i = _fpga_interleaved(rope_q)
-        rope_k_i = _fpga_interleaved(rope_k)
-        attn_core = caps.get("L0_attn_core")
-        if attn_core is not None:
-            attn_core = attn_core.reshape(1, S, NH, HD)
-            attn_core_i = _fpga_interleaved(attn_core)
-            caps["L0_attn_core_fpga"] = attn_core_i
-            caps["L0_attn_out_hm"] = _head_major(attn_core_i)
-
-        caps.update({
-            "L0_q_proj": q_proj_i,
-            "L0_k_proj": k_proj_i,
-            "L0_v_proj": v_proj_i,
-            "L0_q_pre_rope": q_norm_i,
-            "L0_k_pre_rope": k_norm_i,
-            "L0_v_norm": v_norm_i,
-            "L0_rope_q": rope_q_i,
-            "L0_rope_k": rope_k_i,
-            "L0_q_hm": _head_major(rope_q_i, prescale=float(HD) ** 0.5),
-            "L0_k_hm": _head_major(rope_k_i),
-            "L0_v_hm": _head_major(v_norm_i),
-        })
-    except Exception as e:
-        print(f"  [numeric] (layer-0 attention-layout refs unavailable: {type(e).__name__}: {e})")
-
-
-def _rows(t):
-    return t.shape[-2] if (t is not None and hasattr(t, "dim") and t.dim() >= 2) else -1
-
-
-def build_references(cfg, image_path, prompt):
-    """Run the HF vision tower TWICE and capture per-stage outputs (A, B, C):
-
-      * "hf"      — full-precision HF vision tower (ground truth).
-      * "hostref" — same HF tower with its Linear weights round-tripped through IF4
-                    (the FPGA's vision quantization) → mimics the hardware numerics.
-
-    Both are captured in one model load (full-precision forward first, then quantize
-    in place and forward again). Returns (refs, meta) with refs["hf"], refs["hostref"].
-    """
-    hf_model, model_dir = _ensure_hf_model(SCRIPT_DIR, cfg)
+def build_vision_references(ue, image_path, prompt):
+    """Build only B/C references; no patch-embed or layer-0 checkpoints."""
+    hf_model, model_dir = g4r._ensure_hf_model(ue.script_dir, ue._cfg)
     hf_model.eval()
     processor = AutoProcessor.from_pretrained(model_dir)
-
-    # Preprocess identically to _run_vision_encoder_fpga (pixels are prompt-independent).
-    image = Image.open(image_path).convert("RGB").resize(VISION_CANONICAL_SIZE, Image.BICUBIC)
-    conversation = [{"role": "user", "content": [{"type": "image"}, {"type": "text", "text": prompt}]}]
-    text_prompt = processor.apply_chat_template(conversation, add_generation_prompt=True, tokenize=False)
-    inputs = processor(text=[text_prompt], images=[[image]], return_tensors="pt")
+    image = Image.open(image_path).convert("RGB").resize(
+        g4r.VISION_CANONICAL_SIZE, Image.BICUBIC)
+    conversation = [{"role": "user", "content": [
+        {"type": "image"}, {"type": "text", "text": prompt}]}]
+    text = processor.apply_chat_template(
+        conversation, add_generation_prompt=True, tokenize=False)
+    inputs = processor(text=[text], images=[[image]], return_tensors="pt")
     pixel_values = inputs["pixel_values"]
-    pixel_position_ids = inputs["image_position_ids"]
-    padding = (pixel_position_ids == -1).all(dim=-1).squeeze(0)  # [S] True = padding patch
-    n_patches = pixel_values.shape[1]
-
-    vt = hf_model.model.vision_tower
+    position_ids = inputs["image_position_ids"]
+    padding = (position_ids == -1).all(dim=-1).squeeze(0)
+    vision_tower = hf_model.model.vision_tower
 
     def _capture():
-        """One get_image_features forward; return {A, B, C} (B = pre-pool encoder out)
-        plus layer-0 intermediates L0_before/L0_after/L0_end for --layer0."""
-        caps = {"layers": []}
-        hooks = [
-            vt.patch_embedder.register_forward_hook(lambda m, i, o: caps.__setitem__("A", _tensor(o))),
-            vt.encoder.register_forward_hook(lambda m, i, o: caps.__setitem__("B", _tensor(o))),
-        ]
+        checkpoint = {}
+        hook = vision_tower.encoder.register_forward_hook(
+            lambda _m, _i, output: checkpoint.__setitem__(
+                "B", _output_tensor(output).detach().float().cpu()))
         try:
-            for L in vt.encoder.layers:
-                hooks.append(L.register_forward_hook(
-                    lambda m, i, o, _c=caps["layers"]: _c.append(_tensor(o))))
-            L0 = vt.encoder.layers[0]
-            hooks.append(L0.input_layernorm.register_forward_hook(
-                lambda m, i, o: caps.__setitem__("L0_before", _tensor(o))))
-            hooks.append(L0.self_attn.register_forward_hook(
-                lambda m, i, o: caps.__setitem__("L0_after", _tensor(o))))
-            hooks.append(L0.register_forward_hook(
-                lambda m, i, o: caps.__setitem__("L0_end", _tensor(o))))
-            # raw attention-core output = the INPUT to o_proj (pre-hook grabs args[0])
-            hooks.append(L0.self_attn.o_proj.register_forward_pre_hook(
-                lambda m, args: caps.__setitem__("L0_attn_core", args[0] if args else None)))
-        except Exception as e:
-            print(f"  [numeric] (layer-0 hooks unavailable: {type(e).__name__}: {e})")
+            with torch.no_grad():
+                output = hf_model.model.get_image_features(
+                    pixel_values=pixel_values.to(torch.bfloat16),
+                    image_position_ids=position_ids)
+        finally:
+            hook.remove()
+        checkpoint["C"] = getattr(
+            output, "pooler_output", output).detach().float().cpu()
+        return checkpoint
+
+    print("  [numeric] vision reference 1/2: full-precision HF")
+    hf_ref = _capture()
+    count = quantize_vision_tower_(
+        vision_tower, g4r.VISION_QUANT_PRECISION)
+    print(f"  [numeric] vision reference 2/2: {count} IF4 Linear weights")
+    host_ref = _capture()
+    return {"hf": hf_ref, "hostref": host_ref}, {"padding": padding}
+
+
+def build_hf_lm_reference(ue, inputs_embeds, prepared_per_layer_inputs):
+    """Run unmodified HF LM on the exact external tensors consumed by FPGA."""
+    hf_model, _ = g4r._ensure_hf_model(ue.script_dir, ue._cfg)
+    hf_model.eval()
+    language_model = hf_model.model.language_model
+    inputs_embeds = inputs_embeds.unsqueeze(0).to(next(hf_model.parameters()).dtype)
+    prepared_per_layer_inputs = prepared_per_layer_inputs.unsqueeze(0).to(
+        inputs_embeds.dtype)
+
+    layer_out = {}
+    def _capture_last_layer(_module, _args, output):
+        layer_out["hidden"] = _output_tensor(output).detach().float().cpu()
+
+    hook = hf_model.model.language_model.layers[-1].register_forward_hook(
+        _capture_last_layer)
+    original_project = language_model.project_per_layer_inputs
+    language_model.project_per_layer_inputs = (
+        lambda _inputs_embeds, per_layer_inputs: per_layer_inputs)
+    try:
         with torch.no_grad():
-            out = hf_model.model.get_image_features(
-                pixel_values=pixel_values.to(torch.bfloat16),
-                image_position_ids=pixel_position_ids)
-        for h in hooks:
-            h.remove()
-        caps["C"] = getattr(out, "pooler_output", out)
-        if _rows(caps.get("B")) != n_patches and caps["layers"] and _rows(caps["layers"][-1]) == n_patches:
-            caps["B"] = caps["layers"][-1]
-        _add_l0_attention_layout_refs(caps, vt, pixel_position_ids)
-        return caps
+            output = language_model(
+                inputs_embeds=inputs_embeds,
+                per_layer_inputs=prepared_per_layer_inputs,
+                attention_mask=torch.ones(inputs_embeds.shape[:2], dtype=torch.long),
+                use_cache=False,
+                return_dict=True)
+    finally:
+        language_model.project_per_layer_inputs = original_project
+        hook.remove()
 
-    print("  [numeric] reference 1/2: full-precision HF vision tower (hf)")
-    hf_refs = _capture()
+    hidden = layer_out["hidden"]
+    norm = output.last_hidden_state.detach().float().cpu()
+    logits = hf_model.lm_head(output.last_hidden_state)
+    softcap = hf_model.config.text_config.final_logit_softcapping
+    if softcap is not None:
+        logits = torch.tanh(logits / softcap) * softcap
+    return {
+        "hidden": hidden,
+        "norm": norm,
+        "logits": logits.detach().float().cpu(),
+    }
 
-    nq = quantize_vision_tower_(vt, VISION_QUANT_PRECISION)
-    print(f"  [numeric] reference 2/2: HF tower + {nq} Linear weights through "
-          f"{VISION_QUANT_PRECISION.upper()} (hostref = mimics FPGA quantization)")
-    hostref = _capture()
 
-    del hf_model
-    return {"hf": hf_refs, "hostref": hostref}, {"padding": padding}
+def _bf16_region(ue, key, shape, layer_idx=None):
+    """Read one exact BF16 tensor from params.bin (including stored gamma offset)."""
+    off = ue.weight_defs[key]
+    if layer_idx is not None:
+        off += layer_idx * ue.weight_defs["LAYER_WEIGHT_SIZE"]
+    count = 1
+    for dim in shape:
+        count *= dim
+    return torch.frombuffer(
+        ue.weight_bin, dtype=torch.bfloat16, count=count, offset=off).reshape(shape)
 
 
-# ---------------------------------------------------------------------------
-# Driver
-# ---------------------------------------------------------------------------
+def _if4_region(ue, scale_key, data_key, n, k, layer_idx=None):
+    """Dequantize the exact adaptive-IF4 bytes consumed by an FPGA projection."""
+    layer_off = (layer_idx or 0) * ue.weight_defs["LAYER_WEIGHT_SIZE"]
+    scale_off = ue.weight_defs[scale_key] + layer_off
+    data_off = ue.weight_defs[data_key] + layer_off
+    scale_size = n * (k // 64) * 2
+    data_size = n * k // 2
+    scales = bytes(ue.weight_bin[scale_off:scale_off + scale_size])
+    data = bytes(ue.weight_bin[data_off:data_off + data_size])
+    return quant_lib.dequant("if4", data, scales, n, k, block_size=64)
+
+
+def _hw_rms(x, gamma=None):
+    """FPGA RMS convention, including its SRAM BF16 stage boundaries."""
+    x = x.to(torch.bfloat16)
+    # rms_norm_core broadcasts the LALU's reciprocal-RMS result into a BF16
+    # SRAM vector, then performs gamma as a second BF16 elementwise operation.
+    inv_rms = torch.rsqrt(
+        torch.mean(x.float().square(), dim=-1, keepdim=True)).to(torch.bfloat16)
+    out = (x * inv_rms).to(torch.bfloat16)
+    if gamma is not None:
+        out = (out * gamma.to(torch.bfloat16)).to(torch.bfloat16)
+    return out
+
+
+def _hw_linear(x, weight, gelu=False):
+    """Matmul with one BF16 writeback; GELU stays fused before writeback."""
+    out = x.to(torch.bfloat16).float() @ weight.to(torch.bfloat16).float().T
+    if gelu:
+        # LALU_ACT_GELU_B=0xbfd9 encodes -1.6953125; the fused datapath
+        # computes x*sigmoid(1.6953125*x) directly from the matmul accumulator.
+        out = out * torch.sigmoid(1.6953125 * out)
+    return out.to(torch.bfloat16)
+
+
+def _hw_add(a, b):
+    """One FPGA BF16 elementwise-add writeback."""
+    return (a.to(torch.bfloat16) + b.to(torch.bfloat16)).to(torch.bfloat16)
+
+
+def _hw_mul(a, b):
+    """One FPGA BF16 elementwise-multiply writeback."""
+    return (a.to(torch.bfloat16) * b.to(torch.bfloat16)).to(torch.bfloat16)
+
+
+def _hw_rope(x, rope_row, rotary_dim):
+    """Apply the split-half RoPE used by rope_hf_core; retain unrotated tail."""
+    half = rotary_dim // 2
+    cos = rope_row[..., :rotary_dim].float()
+    stored_sin = rope_row[..., rotary_dim:2 * rotary_dim].float()
+    x0 = x[..., :half].float()
+    x1 = x[..., half:rotary_dim].float()
+    rotated = torch.cat((
+        x0 * cos[..., :half] + x1 * stored_sin[..., :half],
+        x1 * cos[..., half:] + x0 * stored_sin[..., half:],
+    ), dim=-1)
+    if rotary_dim < x.shape[-1]:
+        rotated = torch.cat((rotated, x[..., rotary_dim:].float()), dim=-1)
+    return rotated.to(torch.bfloat16)
+
+
+def _hw_attention(q, k, v, sliding_window=None):
+    """Token-space equivalent of GQA duplication + unified_attention_core."""
+    seq_len, groups, _ = q.shape
+    # QK matmul writes BF16 scores before the bias/softmax phase.
+    scores = torch.einsum(
+        "tgd,sd->tgs",
+        q.to(torch.bfloat16).float(),
+        k.to(torch.bfloat16).float()).to(torch.bfloat16)
+    qi = torch.arange(seq_len).view(seq_len, 1)
+    ki = torch.arange(seq_len).view(1, seq_len)
+    valid = ki <= qi
+    if sliding_window is not None:
+        valid &= ki > (qi - sliding_window)
+    scores.masked_fill_(~valid[:, None, :], float("-inf"))
+    probs = torch.softmax(scores.float(), dim=-1).to(torch.bfloat16)
+    return torch.einsum(
+        "tgs,sd->tgd", probs.float(),
+        v.to(torch.bfloat16).float()).to(torch.bfloat16)
+
+
+def build_fpga_mimic_lm_reference(ue, inputs_embeds, prepared_per_layer_inputs):
+    """Reproduce compile_prefill/decode operation-by-operation from params.bin.
+
+    This deliberately does not call the HF language-model forward. Inputs are
+    exact FPGA DRAM readbacks and every projection uses the serialized IF4/BF16
+    weight tensor that the compiled program consumes.
+    """
+    x = inputs_embeds.to(torch.bfloat16).clone()
+    pli = prepared_per_layer_inputs.to(torch.bfloat16)
+    seq_len = x.shape[0]
+    kv_cache = {}
+    num_pos = ue._cfg["special"]["rope"]["num_positions"]
+    rope_local = _bf16_region(
+        ue, "ROPE_LOCAL", (num_pos, ue.head_dim_sliding * 2))
+    global_rotary_dim = int(
+        ue.head_dim * ue._cfg["special"]["rope"]["partial_rotary_factor_global"])
+    rope_global_raw = _bf16_region(
+        ue, "ROPE_GLOBAL", (num_pos, global_rotary_dim * 2))
+
+    for layer_idx in range(ue.LAYER_SIZE):
+        head_dim, q_size, k_size = ue._get_layer_attention_dims(layer_idx)
+        mlp_dim = ue._get_mlp_elements(layer_idx)
+        rotary_dim = ue._get_rope_dims(layer_idx)
+        residual = x
+
+        gamma = _bf16_region(
+            ue, "BLK0_ATTN_NORM_WEIGHT", (ue.vector_length,), layer_idx)
+        pre_norm = _hw_rms(x, gamma)
+        q_w = _if4_region(ue, "BLK0_ATTN_Q_WEIGHT_SCALE",
+                          "BLK0_ATTN_Q_WEIGHT_DATA", q_size,
+                          ue.vector_length, layer_idx)
+        q = _hw_linear(pre_norm, q_w).reshape(seq_len, ue.group_size, head_dim)
+
+        if layer_idx not in ue._kv_shared_map:
+            k_w = _if4_region(ue, "BLK0_ATTN_K_WEIGHT_SCALE",
+                              "BLK0_ATTN_K_WEIGHT_DATA", k_size,
+                              ue.vector_length, layer_idx)
+            v_w = _if4_region(ue, "BLK0_ATTN_V_WEIGHT_SCALE",
+                              "BLK0_ATTN_V_WEIGHT_DATA", k_size,
+                              ue.vector_length, layer_idx)
+            k = _hw_linear(pre_norm, k_w)
+            v = _hw_rms(_hw_linear(pre_norm, v_w))
+            gamma_k = _bf16_region(
+                ue, "BLK0_ATTN_K_NORM_WEIGHT", (ue.head_dim,), layer_idx)[:head_dim]
+            k = _hw_rms(k, gamma_k)
+        gamma_q = _bf16_region(
+            ue, "BLK0_ATTN_Q_NORM_WEIGHT", (ue.head_dim,), layer_idx)[:head_dim]
+        q = _hw_rms(q, gamma_q)
+
+        if layer_idx in ue._rope_global_layers:
+            rope = rope_global_raw[:, :2 * rotary_dim]
+        else:
+            rope = rope_local[:, :2 * rotary_dim]
+        q = _hw_rope(q, rope[:seq_len, None, :], rotary_dim)
+        if layer_idx not in ue._kv_shared_map:
+            k = _hw_rope(k, rope[:seq_len], rotary_dim)
+            kv_cache[layer_idx] = (k, v)
+        else:
+            k, v = kv_cache[ue._kv_shared_map[layer_idx]]
+
+        sliding = None if layer_idx in ue._full_attention_layers else ue.sliding_window
+        attn = _hw_attention(q, k, v, sliding).reshape(seq_len, q_size)
+        o_w = _if4_region(ue, "BLK0_ATTN_OUTPUT_WEIGHT_SCALE",
+                          "BLK0_ATTN_OUTPUT_WEIGHT_DATA", ue.vector_length,
+                          q_size, layer_idx)
+        attn_out = _hw_linear(attn, o_w)
+        gamma = _bf16_region(
+            ue, "BLK0_POST_ATTENTION_NORM_WEIGHT", (ue.vector_length,), layer_idx)
+        x = _hw_add(residual, _hw_rms(attn_out, gamma))
+
+        gamma = _bf16_region(
+            ue, "BLK0_FFN_NORM_WEIGHT", (ue.vector_length,), layer_idx)
+        pre_mlp = _hw_rms(x, gamma)
+        gate_w = _if4_region(ue, "BLK0_FFN_GATE_WEIGHT_SCALE",
+                             "BLK0_FFN_GATE_WEIGHT_DATA", mlp_dim,
+                             ue.vector_length, layer_idx)
+        up_w = _if4_region(ue, "BLK0_FFN_UP_WEIGHT_SCALE",
+                           "BLK0_FFN_UP_WEIGHT_DATA", mlp_dim,
+                           ue.vector_length, layer_idx)
+        gate = _hw_linear(pre_mlp, gate_w, gelu=True)
+        up = _hw_linear(pre_mlp, up_w)
+        mlp_product = _hw_mul(gate, up)
+        del gate_w, up_w, gate, up
+        down_w = _if4_region(ue, "BLK0_FFN_DOWN_WEIGHT_SCALE",
+                             "BLK0_FFN_DOWN_WEIGHT_DATA", ue.vector_length,
+                             mlp_dim, layer_idx)
+        down = _hw_linear(mlp_product, down_w)
+        gamma = _bf16_region(
+            ue, "BLK0_POST_FFW_NORM_WEIGHT", (ue.vector_length,), layer_idx)
+        x = _hw_add(x, _hw_rms(down, gamma))
+
+        gate_w = _bf16_region(
+            ue, "BLK0_PER_LAYER_INPUT_GATE_WEIGHT",
+            (ue.per_layer_input_dim, ue.vector_length), layer_idx)
+        inj_gate = _hw_linear(x, gate_w, gelu=True)
+        inj_gate = _hw_mul(inj_gate, pli[:, layer_idx])
+        proj_w = _bf16_region(
+            ue, "BLK0_PER_LAYER_PROJECTION_WEIGHT",
+            (ue.vector_length, ue.per_layer_input_dim), layer_idx)
+        injection = _hw_linear(inj_gate, proj_w)
+        gamma = _bf16_region(
+            ue, "BLK0_POST_PER_LAYER_INPUT_NORM_WEIGHT",
+            (ue.vector_length,), layer_idx)
+        injection = _hw_rms(injection, gamma)
+        x = _hw_add(x, injection)
+        layer_scale = torch.as_tensor(
+            ue._layer_scalars[layer_idx]).to(torch.bfloat16)
+        x = _hw_mul(x, layer_scale)
+
+    hidden = x.float().cpu()
+    out_gamma = _bf16_region(ue, "OUTPUT_NORM_WEIGHT", (ue.vector_length,))
+    norm = _hw_rms(x, out_gamma)
+    # Prefill never executes LM-head. Compute only the final row, matching the
+    # first decoder token and avoiding a seq_len x 262144 host allocation.
+    lm_head = _if4_region(ue, "LM_HEAD_WEIGHT_SCALE", "LM_HEAD_WEIGHT_DATA",
+                          ue.EMBEDDING_ELEMENTS, ue.vector_length)
+    logits = _hw_linear(norm[-1:], lm_head)
+    return {"hidden": hidden, "norm": norm.float().cpu(),
+            "logits": logits.float().cpu()}
+
+
 def main():
-    parser = argparse.ArgumentParser(
-        description="Gemma4 E2B vision-encoder numeric check: host oracle vs FPGA (SNR).")
-    parser.add_argument("--image", type=str, default=None,
-                        help="Image path (default: shipped yosemite.jpg).")
-    parser.add_argument("--prompt", type=str, default="Describe this image in detail.",
-                        help="Text prompt (vision output is prompt-independent; affects nothing here).")
-    parser.add_argument("--layer0", action="store_true",
-                        help="Deep layer-0 probe: compare FPGA vs host at 3 points inside "
-                             "encoder layer 0 (before-attn / after-attn / end-of-L0). "
-                             "Bakes 3 snapshot DMAs into the layer-0 ISA; forces a bin rebuild.")
-    parser.add_argument("--local-weights", action="store_true")
-    parser.add_argument("--dev", type=str, default="xdma0")
-    parser.add_argument("--cycle", type=float, default=5.62)
-    args = parser.parse_args()
-
-    # Force the FPGA vision path (this branch defaults to the HF host-feature
-    # fallback via GEMMA4_VISION_FPGA); numeric.py needs the FPGA readbacks (_vis_ckpt).
-    os.environ["GEMMA4_VISION_FPGA"] = "1"
-    if args.layer0:
-        os.environ["GEMMA4_VIS_L0_CKPT"] = "1"
-        print("  [numeric] --layer0: layer-0 checkpoint snapshots enabled (forces bin rebuild)")
+    p = argparse.ArgumentParser(
+        description="Gemma4 E2B numeric check: LM prefill/decode, optionally vision.")
+    p.add_argument("--image", type=str, nargs="?", const=g4r.DEFAULT_IMAGE, default=None,
+                   help="Enable VLM checks. Bare --image uses the default image; an optional "
+                        "path or filename selects another image.")
+    p.add_argument("--prompt", type=str, default=None,
+                   help="Text prompt. Without --image, default is the built-in LM test question.")
+    p.add_argument("--dev", type=str, default="xdma0")
+    p.add_argument("--cycle", type=float, default=1000 / 198.3256)
+    args = p.parse_args()
 
     set_dma_device(args.dev)
-    g4.DMA_DEVICE_H2C = user_dma_core.DMA_DEVICE_H2C
-    g4.DMA_DEVICE_C2H = user_dma_core.DMA_DEVICE_C2H
-    g4.DMA_DEVICE_USER = user_dma_core.DMA_DEVICE_USER
+    # The split mixin modules import DMA_DEVICE_* by value, so keep all three
+    # module namespaces synchronized with set_dma_device().
+    for module in (g4r, g4_lm_module, g4_vision_module):
+        module.DMA_DEVICE_H2C = user_dma_core.DMA_DEVICE_H2C
+        module.DMA_DEVICE_C2H = user_dma_core.DMA_DEVICE_C2H
+        module.DMA_DEVICE_USER = user_dma_core.DMA_DEVICE_USER
     user_dma_core.CLOCK_CYCLE_TIME_NS = args.cycle
-    print(f"Using DMA device: {args.dev}")
+    print(f"Using DMA device: {args.dev}  (cycle {args.cycle:.4f} ns)")
 
-    image_path = args.image or DEFAULT_IMAGE
-    if not os.path.exists(image_path):
-        raise SystemExit(f"Image not found: {image_path}")
+    ue = g4r.Gemma4_UnifiedEngine()
 
-    # --- FPGA side: compile (if needed), load, run the vision encoder ---
-    ue = Gemma4_UnifiedEngine(local_weights=args.local_weights)
-    bin_dir = os.path.join(SCRIPT_DIR, "gemma4_e2b_bin")
-    if args.layer0:
-        # --layer0 changes the vision ISA + tensor layout → must rebuild the bin.
-        for _f in ("programs.bin", "programs.json"):
-            try:
-                os.remove(os.path.join(bin_dir, _f))
-            except FileNotFoundError:
-                pass
-    if not (os.path.exists(os.path.join(bin_dir, "programs.bin"))
-            and os.path.exists(os.path.join(bin_dir, "programs.json"))):
-        print("--- Compile: building LM + vision + audio into programs.bin ---")
-        ue.compile_instruction_bin(image_path=DEFAULT_IMAGE, audio_path=DEFAULT_AUDIO)
-    ue.load_instruction_bin()
+    if args.image:
+        # Resolve a bare filename against test_samples/, like test.py.
+        image_path = args.image
+        if not os.path.isfile(image_path):
+            candidate = os.path.join(
+                os.path.dirname(g4r.DEFAULT_IMAGE), os.path.basename(image_path))
+            if os.path.isfile(candidate):
+                image_path = candidate
+        if not os.path.isfile(image_path):
+            raise SystemExit(f"Image not found: {image_path}")
 
-    print(f"\n[numeric] running FPGA vision encoder on {image_path} ...")
-    ue.set_prefill_seq_vlm(image_path, prompt=args.prompt)
-    ckpt = getattr(ue, "_vis_ckpt", None)
-    if not ckpt:
-        raise SystemExit("FPGA checkpoints missing (self._vis_ckpt not set).")
+        # FPGA vision encoder plus the existing vision numeric comparisons.
+        print(f"\n[numeric] running REFACTOR FPGA vision encoder on {image_path} ...")
+        ue.set_prefill_seq_vlm(image_path, args.prompt)
+        ckpt = getattr(ue, "_vis_ckpt", None)
+        if not ckpt:
+            raise SystemExit("FPGA checkpoints missing (self._vis_ckpt not set).")
 
-    # --- Host side: build BOTH references (hostref = IF4/HW-mimicking, hf = full precision) ---
-    print(f"\n[numeric] building references ...")
-    refs, meta = build_references(ue._cfg, image_path, args.prompt)
-    real_rows = ~meta["padding"]  # non-padding patches
+        print("\n[numeric] building vision references ...")
+        vision_prompt = args.prompt or "Describe this image in detail."
+        refs, meta = build_vision_references(ue, image_path, vision_prompt)
+        real = ~meta["padding"]
+        stages = [("B encoder_out", "B", "encoder_out", real),
+                  ("C image_features", "C", "image_features", None)]
 
-    if args.layer0:
-        l0 = getattr(ue, "_vis_l0", None)
-        if not l0:
-            raise SystemExit("layer-0 snapshots missing (self._vis_l0 not set).")
-        def _health(name, t, mask):
-            x = t.detach().float()
-            x = x[mask] if mask is not None and mask.numel() == x.shape[0] else x
-            xf = x[torch.isfinite(x)]
-            rms = float(torch.sqrt((xf * xf).mean())) if xf.numel() else float('nan')
-            print(f"  [numeric] {name:16s} finite={int(torch.isfinite(x).sum())}/{x.numel()} "
-                  f"nan={int(torch.isnan(x).sum())} inf={int(torch.isinf(x).sum())} rms={rms:.4g} "
-                  f"min={float(xf.min()) if xf.numel() else float('nan'):.4g} "
-                  f"max={float(xf.max()) if xf.numel() else float('nan'):.4g}")
-
-        print("\n[numeric] ===== FPGA vs references — LAYER 0 (SNR dB; real patches) =====")
-        S = real_rows.numel()
-        aS = ((S + 63) // 64) * 64
-        nh = l0["q_hm"].shape[0] // aS if "q_hm" in l0 else 0
-        hm_mask = torch.zeros(nh, aS, dtype=torch.bool)
-        if nh:
-            hm_mask[:, :S] = real_rows.unsqueeze(0).expand(nh, S)
-            hm_mask = hm_mask.reshape(nh * aS)
-        def _report_if(name, ref_key, res_key, ref_pack, mask=real_rows):
-            if ref_key in ref_pack and res_key in l0:
-                report(name, ref_pack[ref_key], l0[res_key], row_mask=mask)
-
-        print("  FPGA vs HOSTREF (IF4 — mimics hardware):")
-        report("L0 before-attn", refs["hostref"]["L0_before"], l0["before_attn"], row_mask=real_rows)
-        _report_if("L0 q-proj", "L0_q_proj", "q_proj", refs["hostref"])
-        _report_if("L0 k-proj", "L0_k_proj", "k_proj", refs["hostref"])
-        _report_if("L0 v-proj", "L0_v_proj", "v_proj", refs["hostref"])
-        _report_if("L0 q-pre-rope", "L0_q_pre_rope", "q_pre_rope", refs["hostref"])
-        _report_if("L0 k-pre-rope", "L0_k_pre_rope", "k_pre_rope", refs["hostref"])
-        _report_if("L0 v-norm", "L0_v_norm", "v_norm", refs["hostref"])
-        _report_if("L0 rope-q", "L0_rope_q", "rope_q", refs["hostref"])
-        _report_if("L0 rope-k", "L0_rope_k", "rope_k", refs["hostref"])
-        _report_if("L0 q-hm", "L0_q_hm", "q_hm", refs["hostref"], hm_mask)
-        _report_if("L0 k-hm", "L0_k_hm", "k_hm", refs["hostref"], hm_mask)
-        _report_if("L0 v-hm", "L0_v_hm", "v_hm", refs["hostref"], hm_mask)
-        _report_if("L0 attn-out-hm", "L0_attn_out_hm", "attn_out_hm", refs["hostref"], hm_mask)
-        _report_if("L0 attn-core fpga", "L0_attn_core_fpga", "attn_core", refs["hostref"])
-        report("L0 attn-core",   refs["hostref"]["L0_attn_core"], l0["attn_core"], row_mask=real_rows)
-        report("L0 after-attn",  refs["hostref"]["L0_after"],  l0["after_attn"],  row_mask=real_rows)
-        report("L0 end",         refs["hostref"]["L0_end"],    l0["end_l0"],      row_mask=real_rows)
-        print("  FPGA vs HF (full-precision ground truth):")
-        report("L0 before-attn", refs["hf"]["L0_before"], l0["before_attn"], row_mask=real_rows)
-        _report_if("L0 q-proj", "L0_q_proj", "q_proj", refs["hf"])
-        _report_if("L0 k-proj", "L0_k_proj", "k_proj", refs["hf"])
-        _report_if("L0 v-proj", "L0_v_proj", "v_proj", refs["hf"])
-        _report_if("L0 q-pre-rope", "L0_q_pre_rope", "q_pre_rope", refs["hf"])
-        _report_if("L0 k-pre-rope", "L0_k_pre_rope", "k_pre_rope", refs["hf"])
-        _report_if("L0 v-norm", "L0_v_norm", "v_norm", refs["hf"])
-        _report_if("L0 rope-q", "L0_rope_q", "rope_q", refs["hf"])
-        _report_if("L0 rope-k", "L0_rope_k", "rope_k", refs["hf"])
-        _report_if("L0 q-hm", "L0_q_hm", "q_hm", refs["hf"], hm_mask)
-        _report_if("L0 k-hm", "L0_k_hm", "k_hm", refs["hf"], hm_mask)
-        _report_if("L0 v-hm", "L0_v_hm", "v_hm", refs["hf"], hm_mask)
-        _report_if("L0 attn-out-hm", "L0_attn_out_hm", "attn_out_hm", refs["hf"], hm_mask)
-        _report_if("L0 attn-core fpga", "L0_attn_core_fpga", "attn_core", refs["hf"])
-        report("L0 attn-core",   refs["hf"]["L0_attn_core"], l0["attn_core"], row_mask=real_rows)
-        report("L0 after-attn",  refs["hf"]["L0_after"],  l0["after_attn"],  row_mask=real_rows)
-        report("L0 end",         refs["hf"]["L0_end"],    l0["end_l0"],      row_mask=real_rows)
-        print("  roped-Q health (no clean HF ref; checks RoPE didn't blow up):")
-        _health("L0 rope-Q FPGA", l0["rope_q"], real_rows)
-
-        # RoPE is a rotation → it MUST preserve each row's L2 norm. Compare the
-        # per-row norm of Q before RoPE (q_norm output) vs after RoPE. A ratio far
-        # from 1.0 means the FPGA 2D-RoPE is broken (magnitude not preserved).
-        if "q_pre_rope" in l0:
-            pre = l0["q_pre_rope"].float()[real_rows]
-            post = l0["rope_q"].float()[real_rows]
-            npre = torch.linalg.vector_norm(pre, dim=-1)
-            npost = torch.linalg.vector_norm(post, dim=-1)
-            ratio = (npost / (npre + 1e-9))
-            print(f"  [numeric] RoPE norm-preservation: |post|/|pre| per-row "
-                  f"mean={float(ratio.mean()):.4f} median={float(ratio.median()):.4f} "
-                  f"min={float(ratio.min()):.4f} max={float(ratio.max()):.4f}  "
-                  f"(RoPE is a rotation → must be ≈1.0; rms pre={float(torch.sqrt((pre*pre).mean())):.4g} post={float(torch.sqrt((post*post).mean())):.4g})")
-
-        # Self-contained single-head (L0,H0) check: recompute the attention on the
-        # EXACT Q/K/V the FPGA core read, and SNR-compare against the FPGA's own OUT.
-        # No HF, no head re-assembly — isolates one attention run. If this is high,
-        # the core is correct in context and the mismatch is upstream (rope/QKV) or
-        # in the head-layout comparison; if low, the core itself is wrong here.
-        h0 = getattr(ue, "_vis_l0h0", None)
-        if h0 is not None:
-            import math as _m
-            Q = h0["q"].float(); K = h0["k"].float(); V = h0["v"].float()
-            OUT = h0["out"].float(); BIAS = h0["bias"].float()
-            HD = Q.shape[-1]
-            # Core scales Q by 1/sqrt(HD) internally (Q here is the sqrt(HD)-pre-scaled input).
-            scores = (Q * (1.0 / _m.sqrt(HD))) @ K.t() + BIAS
-            P = torch.softmax(scores, dim=-1)
-            ref = P @ V
-            print("  self-contained head-0 attention (host recompute on FPGA's own Q/K/V):")
-            report("L0H0 attn self", ref, OUT, row_mask=real_rows)
+        print("\n[numeric] ===== FPGA (refactor) vs references — SNR dB (real patches) =====")
+        print("  FPGA vs HOSTREF (IF4 — mimics hardware; high = kernel correct):")
+        for name, rkey, ckey, mask in stages:
+            report(name, refs["hostref"][rkey], ckpt[ckey], row_mask=mask)
+        print("  FPGA vs HF (full-precision ground truth; gap = IF4 quant loss):")
+        for name, rkey, ckey, mask in stages:
+            report(name, refs["hf"][rkey], ckpt[ckey], row_mask=mask)
+    elif args.prompt:
+        print(f"[numeric] LM-only mode, prompt: {args.prompt!r}")
+        ue.set_prefill_seq(args.prompt)
     else:
-        print("\n[numeric] ===== FPGA vs references (SNR dB; higher = closer; real patches) =====")
-        print("  FPGA vs HOSTREF (HF tower with IF4-quantized weights — mimics hardware numerics):")
-        report("A patch_embed", refs["hostref"]["A"], ckpt["A"], row_mask=real_rows)
-        report("B encoder_out", refs["hostref"]["B"], ckpt["B"], row_mask=real_rows)
-        report("C image_feat",  refs["hostref"]["C"], ckpt["C"])
-        print("  FPGA vs HF (full-precision transformers ground truth):")
-        report("A patch_embed", refs["hf"]["A"], ckpt["A"], row_mask=real_rows)
-        report("B encoder_out", refs["hf"]["B"], ckpt["B"], row_mask=real_rows)
-        report("C image_feat",  refs["hf"]["C"], ckpt["C"])
+        print("[numeric] LM-only mode, built-in default prompt")
+        ue.set_prefill_seq()
 
-    print("\n[numeric] done.")
+    ue.compile_gemma4()
+    lm_meta = ue._load_program_section("lm")
+    ue._preamble_addr = ue.get_program_dram_addr()
+    prefill_addr = int(lm_meta["prefill_program_start_addr"], 16)
+    decoder_addr = int(lm_meta["decoder_program_start_addr"], 16)
+
+    print("\n[numeric] running FPGA LM prefill ...")
+    ue.run_prefill(prefill_addr, flops=lm_meta["prefill_total_flops"])
+    prefill_len = len(ue.prefill_seq) - 1
+    prefill_inputs = ue.dma_from_accelerator_memory(
+        ue.LAYER0_INPUT_DRAM,
+        (prefill_len, ue.vector_length)).cpu()
+    prefill_per_layer = ue.dma_from_accelerator_memory(
+        ue.PER_LAYER_INPUTS_DRAM,
+        (prefill_len * ue.LAYER_SIZE, ue.per_layer_input_dim)).reshape(
+            prefill_len, ue.LAYER_SIZE, ue.per_layer_input_dim).cpu()
+    hw_prefill_hidden = ue.dma_from_accelerator_memory(
+        ue.LAYER0_OUTPUT_DRAM,
+        (prefill_len, ue.vector_length)).float().cpu()
+
+    print("\n[numeric] running exactly one FPGA decoder step ...")
+    saved_max_context = ue.MAX_CONTEXT_SIZE
+    ue.MAX_CONTEXT_SIZE = len(ue.prefill_seq)
+    try:
+        ue.run_decoder(
+            [lm_meta["decoder_program_size"]], decoder_addr,
+            token_id=ue.prefill_seq[-1],
+            flops_per_token=[lm_meta["decoder_total_flops"]])
+    finally:
+        ue.MAX_CONTEXT_SIZE = saved_max_context
+
+    decoder_input = ue.dma_from_accelerator_memory(
+        ue.LAYER0_INPUT_DRAM, (1, ue.vector_length)).cpu()
+    decoder_per_layer = ue.dma_from_accelerator_memory(
+        ue.PER_LAYER_INPUTS_DRAM,
+        (ue.LAYER_SIZE, ue.per_layer_input_dim)).reshape(
+            1, ue.LAYER_SIZE, ue.per_layer_input_dim).cpu()
+    hw_decoder_hidden = ue.dma_from_accelerator_memory(
+        ue.LAYER0_OUTPUT_DRAM, (1, ue.vector_length)).float().cpu()
+    hw_decoder_norm = ue.dma_from_accelerator_memory(
+        ue.OUTPUT_NORM_DRAM, (1, ue.vector_length)).float().cpu()
+    hw_decoder_logits = ue.dma_from_accelerator_memory(
+        ue.LOGITS_DRAM, (1, ue.EMBEDDING_ELEMENTS)).float().cpu()
+
+    all_inputs = torch.cat((prefill_inputs, decoder_input), dim=0)
+    all_per_layer = torch.cat((prefill_per_layer, decoder_per_layer), dim=0)
+
+    print("\n[numeric] building full-precision LM references from exact FPGA inputs ...")
+    hf_prefill = build_hf_lm_reference(
+        ue, prefill_inputs, prefill_per_layer)
+    hf_decode = build_hf_lm_reference(ue, all_inputs, all_per_layer)
+
+    print("[numeric] building operation-level FPGA-mimicking LM reference "
+          "from exact params.bin weights ...")
+    mimic_decode = build_fpga_mimic_lm_reference(
+        ue, all_inputs, all_per_layer)
+    mimic_prefill_hidden = mimic_decode["hidden"][:prefill_len]
+
+    print("\n[numeric] ===== LM PREFILL (exact FPGA inputs) =====")
+    print("  FPGA vs HOSTREF (operation-level params.bin oracle):")
+    report("prefill hidden", mimic_prefill_hidden, hw_prefill_hidden)
+    print("  FPGA vs HF (full-precision projections):")
+    report("prefill hidden", hf_prefill["hidden"], hw_prefill_hidden)
+
+    ref_pos = all_inputs.shape[0] - 1
+    mimic_hidden = mimic_decode["hidden"][ref_pos:ref_pos + 1]
+    mimic_norm = mimic_decode["norm"][ref_pos:ref_pos + 1]
+    mimic_logits = mimic_decode["logits"]
+    hf_hidden = hf_decode["hidden"][:, ref_pos]
+    hf_norm = hf_decode["norm"][:, ref_pos]
+    hf_logits = hf_decode["logits"][:, ref_pos]
+
+    print("\n[numeric] ===== LM DECODER (first generated token) =====")
+    print("  FPGA vs HOSTREF (operation-level params.bin oracle):")
+    report("decoder hidden", mimic_hidden, hw_decoder_hidden)
+    report("decoder norm", mimic_norm, hw_decoder_norm)
+    report("decoder logits", mimic_logits, hw_decoder_logits)
+    print("  FPGA vs HF (full-precision projections):")
+    report("decoder hidden", hf_hidden, hw_decoder_hidden)
+    report("decoder norm", hf_norm, hw_decoder_norm)
+    report("decoder logits", hf_logits, hw_decoder_logits)
+    print(f"  argmax: HOSTREF={int(mimic_logits.argmax())} "
+          f"HF={int(hf_logits.argmax())} FPGA={int(hw_decoder_logits.argmax())}")
 
 
 if __name__ == "__main__":
